add the policies for the admin page

+++ app/Http/Controllers/FestivalController.php
@@ -51,6 +51,7 @@ class FestivalController extends Controller
 
     public function delete($festival)
     {
+        $this->authorize('delete', Festival::class);
 
         Festival::find($festival)->delete();
         return redirect()->back();
